Version
In this file, we explained the idea of â€‹â€‹the "HungerStation" application, and identified a set of main requirements that the application should provide. 

Then we conducted a survey to find out the opinions of users about delivery applications, and asked several questions related to the use of these applications, their features, and their importance and impact on society. We found that most users find the applications useful and save time and effort, but there are some drawbacks. 

We also compared similar applications and identified their strengths and weaknesses, explaining the advantages of our application that provides additional services and other features to improve the user experience.


Signed-off-by: 1122325333 <lylayaseen7777@icloud.com>