Version2
I made some important adjustments to improve the arrangement and information provided
Rearrange sections: I arranged some sections logically to improve the flow of information. For example, functional and non-functional requirements were organized separately and in a clear manner to facilitate understanding.
Clarify requirements: I detailed the functional and non-functional requirements to be more clear. Examples were added for each requirement to ensure that the reader understands its importance in the design of the application.
Improved questionnaire analysis: In the questionnaire analysis section, I rephrased the results more clearly, focusing on the most important points such as the advantages and disadvantages mentioned by users, as well as suggestions for improving the application.
Strengthened technical analysis: I detailed the system specifications more, including the technologies used and expected performance, showing how these specifications can contribute to improving the effectiveness of the application.
Error correction: The text was reviewed and grammatical and linguistic errors were corrected, which enhances the quality of writing and makes the content more professional.

Signed-off-by: amjad1425 <alflahi1425@gmail.com>