Latex
In this project, I developed a technical document
The document includes a set of important sections that cover comprehensive information about the application, starting from functional and non-functional requirements, all the way to the architectural design and data. I used LaTeX to write the document, which allowed me to organize the content in a professional and attractive way.
The document included elements
I also used LaTeX features such as tables and images, which helped to better explain the information and enhance understanding.

Signed-off-by: 1122325333 <lylayaseen7777@icloud.com>